Added couple of dues related components

+++ Zeenat/src/main/java/org/anantram/zeenat/repository/AccountRepo.java
@@ -1,8 +1,8 @@
 package org.anantram.zeenat.repository;
 
-import org.anantram.zeenat.domain.Account;
+import org.anantram.zeenat.domain.entities.Account;
 import org.springframework.data.repository.CrudRepository;
 
-public interface AccountRepo extends CrudRepository<Account, Integer> {
+public interface AccountRepo extends CrudRepository<Account, Long> {
 
 }
